Add the “Small Examples” framework
This is an executable, `src/examples/small.ml`, which generates another
executable, `small_examples.ml`, which uses a new module (`Example`) in the
test-library to build a markdown file of examples that are all compiled and if
possible run.

The generated markdown file is used in the build of the documentation.